chore: rename to remarkExtendedMarkdown
