+++ src/components/article/parser/index.tsx
@@ -7,7 +7,7 @@ import remarkGemoji from 'remark-gemoji'
 import remarkGfm from 'remark-gfm'
 import remarkMath from 'remark-math'
 import createMarkdownComponents from './markdownComponents'
-import remarkDiscourse from './remarkDiscourse'
+import remarkExtendedMarkdown from './remarkExtendedMarkdown'
 
 import 'katex/dist/katex.min.css'
 
@@ -29,7 +29,7 @@ const MarkdownContent = async ({ post, translation }: MarkdownContentProps) => {
           [remarkGfm, { singleTilde: false }],
           remarkMath,
           remarkGemoji,
-          remarkDiscourse,
+          remarkExtendedMarkdown,
         ]}
         rehypePlugins={[rehypeRaw, rehypeKatex]}
         components={markdownComponents}
