+++ src/components/article/parser/remarkExtendedMarkdown/details.ts
